Refactor Markdown rendering to use `MarkdownContent` component
